refactor: changed layouts folder structure
