--- layouts/FaqLayout/component/FaqBox/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./FaqBox";
