+++ layouts/HeroLayout/components/Guitar/Guitar.tsx
@@ -1,0 +1,24 @@
+/* eslint-disable @next/next/no-img-element */
+import LightningSVG from "public/static/images/svg/Lightning";
+import LightningRevSVG from "public/static/images/svg/LightningRev";
+
+export default function Guitar() {
+  return (
+    <div className='h-full w-full'>
+      <div className='absolute left-[50%] -bottom-[10%] z-40 h-fit w-[20%] max-w-[100px] -translate-x-[50%] lg:bottom-[15%] lg:left-[47%] lg:w-[35%] lg:max-w-max xl:left-[43%] xl:w-[25%]'>
+        <div className='shadow-animation absolute -bottom-[15%] aspect-square w-full rounded-full'></div>
+        <img
+          className='hover-animation h-full w-full'
+          src='/static/images/guitar_red.png'
+          alt='red guitar'
+        />
+        <div className='shake-animation absolute -bottom-[5%] -left-[100%] w-[65%] lg:-bottom-[15%]'>
+          <LightningRevSVG className='origin-right rotate-[140deg] fill-tertiary-500 ' />
+        </div>
+        <div className='shake-animation absolute bottom-[5%] -right-[70%] w-[40%] lg:-bottom-[10%]'>
+          <LightningSVG className='origin-left rotate-[110deg] fill-tertiary-500' />
+        </div>
+      </div>
+    </div>
+  );
+}
