+++ layouts/HeroLayout/components/Guitar/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Guitar";
