+++ layouts/HeroLayout/components/HeroSlogan/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./HeroSlogan";
