+++ layouts/MainLayout/components/Background/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./Background";
