+++ layouts/MainLayout/components/ContentBox/ContentBox.tsx
