+++ layouts/MainLayout/components/SubtitleBar/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./SubtitleBar";
