+++ layouts/ReportFormLayout/ReportFormLayout.tsx
@@ -1,5 +1,3 @@
-import React from "react";
-
 interface Props {
   children: React.ReactNode;
 }
